Modifica pipeline para evitar sobrescrever banco postgres
Para deixar de usar o campo "processed" do banco postgres, o
pipeline passa a processar itens que foram inserido no banco nas
últimas 24 horas.
