+++ main/__main__.py
@@ -5,7 +5,7 @@ from data_extraction import create_apache_tika_text_extraction
 from database import create_database_interface
 from storage import create_storage_interface
 from index import create_index_interface
-from tasks import extract_text_from_gazettes, extract_themed_excerpts_from_gazettes, get_pending_gazettes, get_themes, rerank_excerpts
+from tasks import embedding_rerank_excerpts, extract_text_from_gazettes, extract_themed_excerpts_from_gazettes, get_gazettes_extracted_since_yesterday, get_themes, tfidf_rerank_excerpts 
 
 
 def is_debug_enabled():
@@ -21,7 +21,7 @@ def enable_debug_if_necessary():
         logging.debug("Debug enabled")
 
 
-def start_to_process_pending_gazettes():
+def daily_processing():
     """
     Setup objects necessary to extract text from the gazettes and call the
     tasks function to extract the text
@@ -31,13 +31,14 @@ def start_to_process_pending_gazettes():
     storage = create_storage_interface()
     index = create_index_interface()
     text_extractor = create_apache_tika_text_extraction()
-    pending_gazettes = get_pending_gazettes(database)
-    text_extracted_gazettes = extract_text_from_gazettes(pending_gazettes, database, storage, index, text_extractor)
+    gazettes_to_be_processed = get_gazettes_extracted_since_yesterday(database)
+    indexed_gazettes = extract_text_from_gazettes(gazettes_to_be_processed, storage, index, text_extractor)
     themes = get_themes()
     for theme in themes:
-        themed_excerpts = extract_themed_excerpts_from_gazettes(theme, text_extracted_gazettes, index)
-        rerank_excerpts(theme, themed_excerpts, index)
+        themed_excerpts = extract_themed_excerpts_from_gazettes(theme, indexed_gazettes, index)
+        embedding_rerank_excerpts(theme, themed_excerpts, index)
+        tfidf_rerank_excerpts(theme, index)
 
 
 if __name__ == "__main__":
-    start_to_process_pending_gazettes()
+    daily_processing()
