+++ tasks/__init__.py
@@ -1,6 +1,7 @@
-from .gazette_excerpts_reranking import rerank_excerpts
-from .gazette_pending_listing import get_pending_gazettes
+from .gazette_excerpts_embedding_reranking import embedding_rerank_excerpts
+from .gazette_excerpts_tfidf_reranking import tfidf_rerank_excerpts
 from .gazette_text_extraction import extract_text_from_gazettes
 from .gazette_themed_excerpts_extraction import extract_themed_excerpts_from_gazettes
 from .gazette_themes_listing import get_themes
 from .interfaces import DatabaseInterface, StorageInterface, IndexInterface, TextExtractorInterface
+from .list_gazettes_to_be_processed import get_gazettes_extracted_since_yesterday
